@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "novem"
-version = "0.4.15"  # do not touch manually, use `poetry run bump_version`
+version = "0.5.0"  # do not touch manually, use `poetry run bump_version`
 description = "python library for the novem data visualisation platform"
 authors = ["Sondov Engen <sondov@novem.no>", "Bjørnar Snoksrud <bjornars@novem.no>", "Martin Myrseth <myme@novem.no>"]
 homepage = "https://novem.io"
@@ -9,16 +9,15 @@ readme = "README.md"
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.8.1"
+python = "^3.10.0"
 requests = "^2.26.0"
 pyreadline3 = "^3.4.1"
 colorama = "^0.4.6"
-urllib3 = "^1.26.15"
 typing-extensions = "^4.9.0"
 
-[tool.poetry.scripts]
-novem = 'novem.cli:run_cli'
-bump_version = 'scripts.bump_version:main'
+# [tool.poetry.scripts]
+# novem = 'novem.cli:run_cli'
+# bump_version = 'scripts.bump_version:main'
 
 [tool.poetry.group.dev.dependencies]
 flake8 = "^7.1.1"
@@ -32,8 +31,8 @@ pytest-cov = "^5.0.0"
 requests-mock = "^1.12.1"
 types-requests = "^2.27.7"
 black = "^24.8.0"
-pandas = { version = "2.2.2", python = "^3.10" }
-pandas-stubs = { version = "2.2.2.240807" , python = "^3.10" }
+pandas = "2.2.2"
+pandas-stubs = "2.2.2.240807"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -41,25 +40,11 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 120
-target-version = ['py38']
+target-version = ['py310']
 include = '\.pyi?$'
 exclude = '''
-
 (
-  /(
-      \.eggs         # exclude a few common directories in the
-    | \.git          # root of the project
-    | \.hg
-    | \.mypy_cache
-    | \.tox
-    | \.venv
-    | _build
-    | buck-out
-    | build
-    | dist
-  )/
-  | foo.py           # also separately exclude a file named foo.py in
-                     # the root of the project
+  /( \.git )/
 )
 '''
 
